backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-socketio==5.10.0
python-multipart==0.0.7
pydantic==2.10.2
pydantic-settings==2.2.1
email-validator==2.1.1
httpx==0.26.0
supabase==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
google-generativeai==0.5.4
aiofiles==24.1.0
